UPDATED: -Quote requests now calculates suggested price -The same VM is used for both quote requests and quotes -Changed height, width and length in custom product  from decimal to int -Quote backend to work with changes to quote entity -Quote statuses in seed data to work with new quote logic

+++ BOX API/Models/Custom_Product.cs
@@ -13,9 +13,9 @@ namespace BOX.Models
         [ForeignKey("Cost_Price_Formula_Variables")]
         public int FormulaID { get; set; }
         public virtual Cost_Price_Formula_Variables Cost_Price_Formula_Variables { get; set; }
-        [Required] public decimal Width { get; set; }
-        [Required] public decimal Height { get; set; }
-        [Required] public decimal Length { get; set; }
+        [Required] public int Width { get; set; }
+        [Required] public int Height { get; set; }
+        [Required] public int Length { get; set; }
         [Required] public byte[] Label { get; set; } //what should be printed
         [Required] public int Sides { get; set; } //number of sides to print on
